requirements.txt
flask
pillow
qrcode
cloudconvert